agent-platform/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
firebase-admin
qdrant-client
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
python-dotenv
httpx
pytest
